FindPETSc.cmake: use compilation test to determine minimal linking
With shared libraries, this will only include the interface (petsc*.so)
libs which is optimal because it leads to smaller binaries and makes it
possible to change/update the libs (if soname is stable) without
relinking executables.

With static libraries, it is necessary to explicitly link the entire
dependency graph of PETSc.  We determine which case using a test
executable.

Signed-off-by: Jed Brown <jed@59A2.org>