Add MULTIPASS_C_SOURCE_RUNS() which forces re-running the test.
This is useful when the environment changes.  If we had a way to
guarantee that the environment described in INCLUDES and LIBRARIES was
really a complete description, we could avoid re-running these tests.
In reality, it probably doesn't matter because we don't request the test
to be rerun unless the environment is likely to have changed.

Signed-off-by: Jed Brown <jed@59A2.org>